move conversion trait implementation into impl
